+++ src/main/java/com/kodilla/ecommercee/dto/CartDto.java
@@ -1,20 +1,16 @@
-package com.kodilla.ecommercee;
-
+package com.kodilla.ecommercee.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
-public class Cart {
+public class CartDto {
 
     @Id
     @GeneratedValue
