+++ src/main/java/com/kodilla/ecommercee/entity/Cart.java
@@ -1,16 +1,20 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.entity;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class CartDto {
+@Entity
+public class Cart {
 
     @Id
     @GeneratedValue
