+++ src/main/java/com/kodilla/ecommercee/entity/UserOrder.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.entity;
 
 
 import lombok.AllArgsConstructor;
