+++ src/main/java/com/kodilla/ecommercee/mapper/OrderMapper.java
@@ -1,4 +1,6 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.mapper;
+import com.kodilla.ecommercee.entity.UserOrder;
+import com.kodilla.ecommercee.dto.OrderDto;
 import org.springframework.stereotype.Component;
 import java.util.List;
 import java.util.stream.Collectors;
