+++ src/main/java/com/kodilla/ecommercee/repository/CartRepository.java
@@ -1,4 +1,5 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.repository;
+import com.kodilla.ecommercee.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartRepository extends JpaRepository<Cart,Long> {
