+++ src/main/java/com/kodilla/ecommercee/repository/OrderRepository.java
@@ -1,4 +1,5 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.repository;
+import com.kodilla.ecommercee.entity.UserOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<UserOrder, Long> {
