+++ src/main/java/com/kodilla/ecommercee/service/CartService.java
@@ -1,5 +1,9 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.service;
 
+import com.kodilla.ecommercee.GenericEntity;
+import com.kodilla.ecommercee.entity.UserOrder;
+import com.kodilla.ecommercee.entity.Cart;
+import com.kodilla.ecommercee.repository.CartRepository;
 import org.springframework.stereotype.Service;
 
 @Service
